컨트롤러 통합테스트  (#53)

+++ src/main/java/kr/hhplus/be/server/coupon/presentation/dto/CouponsResponseDTO.java
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.coupon.presentation.dto;
 public record CouponsResponseDTO(
         long couponId,
         String couponName,
-        long couponQuantity
+        int couponQuantity
 
 ) {
 }
